Pull request fixes. Moved route check to Handler to render method. Implemented check for middleware (web). Removed route check in "app.php".
